Remove env varaible
Not only does this not work, if it did, it would hardcode each instance of the service to connect to one container of the api service. Meaning that all requests would be routed to that one container and not leverage the DNS + Scalability of docker applications.

For such a small application, this isn't a big deal. But for learning purposes, best practicies should be applied.