Remove "pop()" during generate
The last array does not need to be removed as it will either be equal in size (e.g. 3 players, 5 races each, 15 selected == 3 equal partisions) OR the last partision not being used (e.g. 2 player, 8 races each, 17 selected == last one not used).

This is beacuse partisions are distributed in order from 0 to 8.